requirements.txt
pynput>=1.7
pywin32>=306
pystray>=0.19
Pillow>=10.0
